+++ Source/Frontend/UI/Extensions/BitmapExtensions.cs
@@ -1,0 +1,20 @@
+namespace RTCV.UI.Extensions
+{
+    using System.Drawing;
+
+    public static class BitmapExtensions
+    {
+        public static void Tint(this Bitmap bmp, Color col)
+        {
+            var rectSize = new Rectangle(0, 0, bmp.Width, bmp.Height);
+
+            using (var g = Graphics.FromImage(bmp))
+            {
+                g.DrawImage(bmp, rectSize);
+
+                var darkBrush = new SolidBrush(col);
+                g.FillRectangle(darkBrush, rectSize);
+            }
+        }
+    }
+}
