+++ Source/Frontend/UI/UI.csproj
@@ -378,6 +378,7 @@
     </Compile>
     <Compile Include="GameProtection.cs" />
     <Compile Include="SortableBindingList.cs" />
+    <Compile Include="Extensions\BitmapExtensions.cs" />
     <Compile Include="Extensions\ControlExtensions.cs" />
     <Compile Include="Extensions\NumberExtensions.cs" />
     <Compile Include="Input\Bindings.cs" />
